% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trace.R
\name{read_trace_files}
\alias{read_trace_files}
\title{Read Trace Files}
\usage{
read_trace_files(
  dir = NULL,
  param_nm_set = NULL,
  iter_set = NULL,
  chain_set = NULL,
 
    pattern = "^\%\%PARAM\%\%\\\\Q_chain_\\\\E\%\%CHAIN\%\%\\\\Q_iter_\\\\E\%\%ITER\%\%\\\\Q.rds\\\\E$"
)
}
\arguments{
\item{dir}{\verb{[NULL, character]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: directory is taken as the newest in \verb{tracer/}
\item \code{character}: read files from this dir
}}

\item{param_nm_set}{\verb{[NULL, character]} (optional, default \code{NULL})

set of parameter names the traces of which will be read into R.
\itemize{
\item \code{NULL}: guess the whole set of names based on file names in \code{dir}
\item \code{character}: only read these parameters
}}

\item{iter_set}{\verb{[NULL, integer]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: read all possible iterations
\item \code{integer}: read these iterations (only)
}}

\item{chain_set}{\verb{[NULL, integer]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: read all possible chains
\item \code{integer}: read these chains (only)
}}

\item{pattern}{\verb{[character]} (mandatory)

default \code{"^\%\%PARAM\%\%\\\\Q_chain_\\\\E\%\%CHAIN\%\%\\\\Q_iter_\\\\E\%\%ITER\%\%\\\\Q.rds\\\\E$"};
determines how iteration, chain, and parameter name are identified from
file names}
}
\description{
Read trace files into list of arrays / data.tables.
}
